feat: Room 클래스 분리
