+++ backend/bang-ggood/src/main/java/com/bang_ggood/checklist/dto/request/ChecklistRequest.java
@@ -4,8 +4,6 @@ import com.bang_ggood.checklist.domain.Checklist;
 import com.bang_ggood.contract.domain.OccupancyMonth;
 import com.bang_ggood.contract.domain.OccupancyPeriod;
 import com.bang_ggood.question.dto.request.QuestionRequest;
-import com.bang_ggood.room.domain.Room;
-import com.bang_ggood.room.dto.request.RoomRequest;
 import com.bang_ggood.user.domain.User;
 import jakarta.validation.Valid;
 import java.util.List;
