+++ backend/bang-ggood/src/main/java/com/bang_ggood/checklist/dto/response/SelectedRoomResponse.java
@@ -1,4 +1,4 @@
-package com.bang_ggood.room.dto.response;
+package com.bang_ggood.checklist.dto.response;
 
 import com.bang_ggood.checklist.domain.Checklist;
 import java.time.LocalDateTime;
