+++ backend/bang-ggood/src/main/java/com/bang_ggood/checklist/dto/response/SelectedRoomResponseV1.java
@@ -1,4 +1,4 @@
-package com.bang_ggood.room.dto.response;
+package com.bang_ggood.checklist.dto.response;
 
 import java.time.LocalDateTime;
 import java.util.List;
