+++ backend/bang-ggood/src/main/java/com/bang_ggood/checklist/service/ChecklistManageService.java
@@ -33,10 +33,8 @@ import com.bang_ggood.question.dto.response.SelectedCategoryQuestionsResponse;
 import com.bang_ggood.question.dto.response.SelectedQuestionResponse;
 import com.bang_ggood.question.service.ChecklistQuestionService;
 import com.bang_ggood.question.service.QuestionService;
-import com.bang_ggood.room.domain.Room;
-import com.bang_ggood.room.dto.request.RoomRequest;
-import com.bang_ggood.room.dto.response.SelectedRoomResponse;
-import com.bang_ggood.room.service.RoomService;
+import com.bang_ggood.checklist.dto.request.RoomRequest;
+import com.bang_ggood.checklist.dto.response.SelectedRoomResponse;
 import com.bang_ggood.station.domain.ChecklistStation;
 import com.bang_ggood.station.dto.response.SubwayStationResponse;
 import com.bang_ggood.station.dto.response.SubwayStationResponses;
