+++ backend/bang-ggood/src/test/java/com/bang_ggood/checklist/repository/ChecklistImageRepositoryTest.java
@@ -9,8 +9,7 @@ import com.bang_ggood.checklist.service.ChecklistService;
 import com.bang_ggood.global.exception.BangggoodException;
 import com.bang_ggood.global.exception.ExceptionCode;
 import com.bang_ggood.room.RoomFixture;
-import com.bang_ggood.room.domain.Room;
-import com.bang_ggood.room.service.RoomService;
+import com.bang_ggood.checklist.service.RoomService;
 import com.bang_ggood.user.UserFixture;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
