+++ backend/bang-ggood/src/test/java/com/bang_ggood/like/controller/ChecklistLikeE2ETest.java
@@ -8,8 +8,7 @@ import com.bang_ggood.checklist.service.ChecklistManageService;
 import com.bang_ggood.like.repository.ChecklistLikeRepository;
 import com.bang_ggood.like.service.ChecklistLikeManageService;
 import com.bang_ggood.room.RoomFixture;
-import com.bang_ggood.room.domain.Room;
-import com.bang_ggood.room.repository.RoomRepository;
+import com.bang_ggood.checklist.repository.RoomRepository;
 import io.restassured.RestAssured;
 import io.restassured.http.ContentType;
 import org.junit.jupiter.api.DisplayName;
