+++ backend/bang-ggood/src/test/java/com/bang_ggood/like/service/ChecklistLikeServiceTest.java
@@ -7,8 +7,7 @@ import com.bang_ggood.checklist.repository.ChecklistRepository;
 import com.bang_ggood.like.domain.ChecklistLike;
 import com.bang_ggood.like.repository.ChecklistLikeRepository;
 import com.bang_ggood.room.RoomFixture;
-import com.bang_ggood.room.domain.Room;
-import com.bang_ggood.room.repository.RoomRepository;
+import com.bang_ggood.checklist.repository.RoomRepository;
 import com.bang_ggood.user.UserFixture;
 import com.bang_ggood.user.domain.User;
 import com.bang_ggood.user.repository.UserRepository;
