+++ backend/bang-ggood/src/test/java/com/bang_ggood/maintenance/repository/ChecklistMaintenanceRepositoryTest.java
@@ -6,8 +6,7 @@ import com.bang_ggood.checklist.domain.Checklist;
 import com.bang_ggood.checklist.repository.ChecklistRepository;
 import com.bang_ggood.maintenance.domain.ChecklistMaintenance;
 import com.bang_ggood.room.RoomFixture;
-import com.bang_ggood.room.domain.Room;
-import com.bang_ggood.room.repository.RoomRepository;
+import com.bang_ggood.checklist.repository.RoomRepository;
 import com.bang_ggood.user.UserFixture;
 import com.bang_ggood.user.domain.User;
 import com.bang_ggood.user.repository.UserRepository;
