+++ backend/bang-ggood/src/test/java/com/bang_ggood/room/RoomFixture.java
@@ -2,10 +2,9 @@ package com.bang_ggood.room;
 
 import com.bang_ggood.contract.domain.OccupancyMonth;
 import com.bang_ggood.contract.domain.OccupancyPeriod;
-import com.bang_ggood.room.domain.FloorLevel;
-import com.bang_ggood.room.domain.Room;
-import com.bang_ggood.room.domain.Structure;
-import com.bang_ggood.room.dto.request.RoomRequest;
+import com.bang_ggood.checklist.domain.FloorLevel;
+import com.bang_ggood.checklist.domain.Structure;
+import com.bang_ggood.checklist.dto.request.RoomRequest;
 import java.util.List;
 
 public class RoomFixture {
