+++ backend/bang-ggood/src/test/java/com/bang_ggood/room/domain/StructureTest.java
@@ -1,5 +1,6 @@
 package com.bang_ggood.room.domain;
 
+import com.bang_ggood.checklist.domain.Structure;
 import com.bang_ggood.global.exception.BangggoodException;
 import com.bang_ggood.global.exception.ExceptionCode;
 import org.junit.jupiter.api.DisplayName;
