+++ backend/bang-ggood/src/test/java/com/bang_ggood/room/service/RoomServiceTest.java
@@ -1,9 +1,9 @@
 package com.bang_ggood.room.service;
 
 import com.bang_ggood.IntegrationTestSupport;
+import com.bang_ggood.checklist.service.RoomService;
 import com.bang_ggood.room.RoomFixture;
-import com.bang_ggood.room.domain.Room;
-import com.bang_ggood.room.repository.RoomRepository;
+import com.bang_ggood.checklist.repository.RoomRepository;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
