+++ backend/bang-ggood/src/test/java/com/bang_ggood/station/service/ChecklistStationServiceTest.java
@@ -5,8 +5,7 @@ import com.bang_ggood.checklist.ChecklistFixture;
 import com.bang_ggood.checklist.domain.Checklist;
 import com.bang_ggood.checklist.repository.ChecklistRepository;
 import com.bang_ggood.room.RoomFixture;
-import com.bang_ggood.room.domain.Room;
-import com.bang_ggood.room.repository.RoomRepository;
+import com.bang_ggood.checklist.repository.RoomRepository;
 import com.bang_ggood.station.domain.ChecklistStation;
 import com.bang_ggood.station.repository.ChecklistStationRepository;
 import com.bang_ggood.user.UserFixture;
